requirements.txt
asgiref==3.4.1
Django==4.0.1
django-parler==2.3
django-translated-fields==0.11.3
djangorestframework==3.13.1
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
